Simplify color update
Instead of subscribing to signals indicating color theme change and then
update colors it is better to simply update the colors at each
snapshot/redraw because the cost of looking up colors is low. This also
makes sure that colors are always consistent with system theme even in
case of signalling issues.